commiting sample code - don't want to stash

+++ Sample/DigitNet/TrainingData.cs
@@ -30,6 +30,6 @@ namespace DigitNet
         /// <value>
         /// The image data for a single digit.
         /// </value>
-        public byte[] Data { get; set; }
+        public double[] Data { get; set; }
     }
 }
